backend/requirements.txt
daphne==2.5.0
fastapi==0.61.1
google-cloud-firestore==1.9.0
google-cloud-pubsub==2.1.0
pydantic==1.6.1
Twisted[tls,http2]==20.3.0
